requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
sqlalchemy>=2.0.25
requests>=2.31.0
python-dotenv>=1.0.0
feedparser>=6.0.10
schedule>=1.2.1
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0
google-api-python-client>=2.110.0
google-auth>=2.25.0
psutil>=5.9.0